Powerups

+++ Assets/Personnel/Bez/Script/RotatingObject.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class RotatingObject : MonoBehaviour
 {
 	[SerializeField]
-	private int rotationspeed;
+	private float rotationspeed;
 
 	private void FixedUpdate() {
 		transform.Rotate(0, rotationspeed, 0, Space.World);
